chore: update column autocomplete
