+++ dashboard/src/features/orgs/projects/database/dataGrid/components/RuleGroupEditor/RuleValueInput.tsx
@@ -1,11 +1,7 @@
-import { ControlledSelect } from '@/components/form/ControlledSelect';
-
 import { Check, ChevronsUpDown } from 'lucide-react';
 
-import { ReadOnlyToggle } from '@/components/presentational/ReadOnlyToggle';
 import type { InputProps } from '@/components/ui/v2/Input';
 import { inputClasses } from '@/components/ui/v2/Input';
-import { Option } from '@/components/ui/v2/Option';
 import { Button } from '@/components/ui/v3/button';
 import {
   Command,
@@ -22,39 +18,26 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from '@/components/ui/v3/popover';
-import type { ColumnAutocompleteProps } from '@/features/orgs/projects/database/dataGrid/components/ColumnAutocomplete';
-import { ColumnAutocomplete } from '@/features/orgs/projects/database/dataGrid/components/ColumnAutocomplete';
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from '@/components/ui/v3/select';
+import {
+  ColumnAutocomplete,
+  type ColumnAutocompleteProps,
+} from '@/features/orgs/projects/database/dataGrid/components/ColumnAutocomplete';
 import type { HasuraOperator } from '@/features/orgs/projects/database/dataGrid/types/dataBrowser';
 import { useProject } from '@/features/orgs/projects/hooks/useProject';
 import { getAllPermissionVariables } from '@/features/projects/permissions/settings/utils/getAllPermissionVariables';
 import { cn } from '@/lib/utils';
 import { useGetRolesPermissionsQuery } from '@/utils/__generated__/graphql';
+import { CommandLoading } from 'cmdk';
 import { useState } from 'react';
 import { useController, useFormContext, useWatch } from 'react-hook-form';
 import useRuleGroupEditor from './useRuleGroupEditor';
-
-const frameworks = [
-  {
-    value: 'next.js',
-    label: 'Next.js',
-  },
-  {
-    value: 'sveltekit',
-    label: 'SvelteKit',
-  },
-  {
-    value: 'nuxt.js',
-    label: 'Nuxt.js',
-  },
-  {
-    value: 'remix',
-    label: 'Remix',
-  },
-  {
-    value: 'astro',
-    label: 'Astro',
-  },
-];
 
 function ColumnSelectorInput({
   name,
@@ -98,7 +81,7 @@ function ColumnSelectorInput({
             : undefined,
         },
       }}
-      onChange={(_event, { value }) => {
+      onChange={({ value }) => {
         if (selectedTablePath === `${schema}.${table}`) {
           setValue(name, [value], { shouldDirty: true });
           return;
@@ -146,21 +129,9 @@ export default function RuleValueInput({
   const { setValue } = useFormContext();
 
   const [open, setOpen] = useState(false);
-  const [comboboxValue, setComboboxValue] = useState('');
   const inputName = `${name}.value`;
-  console.log('error', helperText);
+  const comboboxValue = useWatch({ name: inputName });
   const operator: HasuraOperator = useWatch({ name: `${name}.operator` });
-  const sharedInputSx: InputProps['sx'] = !disabled
-    ? {
-        backgroundColor: (theme) =>
-          theme.palette.mode === 'dark'
-            ? theme.palette.grey[300]
-            : theme.palette.common.white,
-        [`& .${inputClasses.input}`]: {
-          backgroundColor: 'transparent',
-        },
-      }
-    : undefined;
 
   const {
     data,
@@ -173,41 +144,22 @@ export default function RuleValueInput({
 
   if (operator === '_is_null') {
     return (
-      <ControlledSelect
+      <Select
         disabled={disabled}
         name={inputName}
-        fullWidth
-        slotProps={{
-          root: {
-            className: 'lg:!rounded-none h-10',
-            sx: !disabled
-              ? {
-                  backgroundColor: (theme) =>
-                    theme.palette.mode === 'dark'
-                      ? `${theme.palette.grey[300]} !important`
-                      : `${theme.palette.common.white} !important`,
-                }
-              : null,
-          },
-          popper: { disablePortal: false, className: 'z-[10000]' },
+        onValueChange={(newValue: string) => {
+          setValue(inputName, newValue, { shouldDirty: true });
         }}
-        error={error}
-        helperText={helperText}
+        defaultValue={comboboxValue}
       >
-        <Option value="true">
-          <ReadOnlyToggle
-            checked
-            slotProps={{ label: { className: '!text-sm' } }}
-          />
-        </Option>
-
-        <Option value="false">
-          <ReadOnlyToggle
-            checked={false}
-            slotProps={{ label: { className: '!text-sm' } }}
-          />
-        </Option>
-      </ControlledSelect>
+        <SelectTrigger>
+          <SelectValue placeholder="Is null?" />
+        </SelectTrigger>
+        <SelectContent>
+          <SelectItem value="true">true</SelectItem>
+          <SelectItem value="false">false</SelectItem>
+        </SelectContent>
+      </Select>
     );
   }
   const availableHasuraPermissionVariables = getAllPermissionVariables(
@@ -249,16 +201,11 @@ export default function RuleValueInput({
     );
   }
 
-  const selectedFramework = frameworks.find(
-    (framework) => framework.value === comboboxValue,
+  const selectedVariable = availableHasuraPermissionVariables.find(
+    (variable) => variable.value === comboboxValue,
   );
   const comboboxLabel =
-    selectedFramework?.label || comboboxValue || 'Select framework...';
-
-  console.log(
-    'availableHasuraPermissionVariables',
-    availableHasuraPermissionVariables,
-  );
+    selectedVariable?.label || comboboxValue || 'Select variable...';
 
   return (
     <Popover open={open} onOpenChange={setOpen}>
@@ -267,50 +214,25 @@ export default function RuleValueInput({
           variant="outline"
           role="combobox"
           aria-expanded={open}
-          className="w-[200px] justify-between"
+          className="justify-between rounded-l-none rounded-r-none"
         >
           {comboboxLabel}
           <ChevronsUpDown className="opacity-50" />
         </Button>
       </PopoverTrigger>
-      <PopoverContent className="w-[200px] p-0">
-        <Command
-          filter={(item, query) => {
-            console.log('item', item);
-            if (item === '+create+') {
-              console.log('CREATE*++');
-              // Don't show create option if it matches an existing option
-              if (
-                frameworks.some(
-                  (framework) =>
-                    framework.value.toLowerCase() === query.toLowerCase() ||
-                    framework.label.toLowerCase() === query.toLowerCase(),
-                )
-              ) {
-                console.log('RETURNING 0 for query', query);
-                return 0;
-              }
-              return 1;
-            }
-            if (item.toLowerCase().includes(query.trim().toLowerCase())) {
-              return 1;
-            }
-            return 0;
-          }}
-        >
-          <CommandInput placeholder="Search variable..." />
+      <PopoverContent className="p-0">
+        <Command>
+          <CommandInput placeholder="Choose variable..." />
           <CommandList>
             <CommandEmpty>No variable found.</CommandEmpty>
+            {loading && <CommandLoading>Loading...</CommandLoading>}
             <CommandGroup>
               {availableHasuraPermissionVariables.map((variable) => (
                 <CommandItem
                   key={variable.value}
                   value={variable.value}
                   onSelect={(currentValue) => {
-                    console.log('currentValue', currentValue);
-                    setComboboxValue(
-                      currentValue === comboboxValue ? '' : currentValue,
-                    );
+                    setValue(inputName, currentValue, { shouldDirty: true });
                     setOpen(false);
                   }}
                 >
@@ -328,7 +250,7 @@ export default function RuleValueInput({
             </CommandGroup>
             <CommandCreateItem
               onCreate={(currentValue) => {
-                setComboboxValue(currentValue);
+                setValue(inputName, currentValue, { shouldDirty: true });
                 setOpen(false);
               }}
             />
@@ -336,45 +258,5 @@ export default function RuleValueInput({
         </Command>
       </PopoverContent>
     </Popover>
-    // <ControlledAutocomplete
-    //   disabled={disabled}
-    //   freeSolo
-    //   isOptionEqualToValue={(
-    //     option,
-    //     value: string | number | AutocompleteOption<string>,
-    //   ) => {
-    //     if (typeof value !== 'object') {
-    //       return option.value.toLowerCase() === value?.toString().toLowerCase();
-    //     }
-
-    //     return option.value.toLowerCase() === value.value.toLowerCase();
-    //   }}
-    //   name={inputName}
-    //   groupBy={(option) => option.group}
-    //   slotProps={{
-    //     input: {
-    //       className: 'lg:!rounded-none',
-    //       sx: sharedInputSx,
-    //     },
-    //     formControl: { className: '!bg-transparent' },
-    //     paper: { className: 'empty:border-transparent' },
-    //   }}
-    //   fullWidth
-    //   loading={loading}
-    //   loadingText={<ActivityIndicator label="Loading..." />}
-    //   error={Boolean(customClaimsError) || error}
-    //   helperText={customClaimsError?.message || helperText}
-    //   options={availableHasuraPermissionVariables}
-    //   onChange={(_event, _value, reason, details) => {
-    //     if (
-    //       reason !== 'selectOption' &&
-    //       details.option.value !== 'X-Hasura-User-Id'
-    //     ) {
-    //       return;
-    //     }
-
-    //     setValue(inputName, details.option.value, { shouldDirty: true });
-    //   }}
-    // />
   );
 }
